feat : ログをユーザーごとに紐づける

+++ lib/supabase/logs.ts
@@ -1,13 +1,18 @@
 import { supabase } from "../supabase";
 
-export const fetchLogs = async () => {
+export const fetchLogs = async (userId: string) => {
   return await supabase
     .from("logs")
     .select("*")
+    .eq("user_id", userId)
     .order("date", { ascending: false });
 };
 
-export const createLog = async (text: string, tags: string[]) => {
+export const createLog = async (
+  userId: string,
+  text: string,
+  tags: string[],
+) => {
   return await supabase
     .from("logs")
     .insert([
@@ -15,13 +20,19 @@ export const createLog = async (text: string, tags: string[]) => {
         text,
         date: new Date().toISOString(),
         tags,
+        user_id: userId,
       },
     ])
     .select()
     .single();
 };
 
-export const updateLog = async (id: string, text: string, tags: string[]) => {
+export const updateLog = async (
+  userId: string,
+  id: string,
+  text: string,
+  tags: string[],
+) => {
   return await supabase
     .from("logs")
     .update({
@@ -29,10 +40,11 @@ export const updateLog = async (id: string, text: string, tags: string[]) => {
       tags,
     })
     .eq("id", id)
+    .eq("user_id", userId)
     .select()
     .single();
 };
 
-export const removeLog = async (id: string) => {
-  return await supabase.from("logs").delete().eq("id", id);
+export const removeLog = async (userId: string, id: string) => {
+  return await supabase.from("logs").delete().eq("id", id).eq("user_id", userId);
 };
